recipe will not run if brew vim & macvim installed
This recipe now follows the convention of our other recipes
(e.g. if application is already installed, don't reinstall).

This is especially important with the vim build which is
currently fragile, and liable to abort the entire chef run:
Don't rebuild if we don't need to.

Users who are sure they want to rebuild vim can force a
rebuild by either removing /usr/local/bin/vim or
/Applications/MacVim.app.