Use system ruby in .rvmrc
pivotal_workstation should not rely on a more modern Ruby verion than
the one that ships with OS X Mountain Lion. Using the system ruby seems
to be the easiest way to acheive this. We do lose gemset support
with this change, though it isn't clear to me that we need this.